Fix grid issues on landing view
