+++ package.json
@@ -14,6 +14,7 @@
     "@fontsource/roboto": "^4.3.0",
     "add": "^2.0.6",
     "emotion": "^11.0.0",
+    "facepaint": "^1.2.1",
     "gray-matter": "^4.0.3",
     "next": "10.2.0",
     "react": "17.0.2",
